gather some statistics from db and bugfixes
statistics for rating dream mood, clarity and sleep quality are now gathered from the database. Fixed the bug that cancelling the editing of an audio entry would delete the audio

+++ app/src/main/java/com/bitflaker/lucidsourcekit/main/createjournalentry/AddTextEntry.java
@@ -321,7 +321,7 @@ public class AddTextEntry extends AppCompatActivity {
 
     @Override
     protected void onStop() {
-        if(!storedByUser){
+        if(!storedByUser && !isInEditingMode){
             for (String recordedAudio : recordedAudios) {
                 File audio = new File(getFilesDir().getAbsolutePath() + recordedAudio);
                 audio.delete();
